sdk-py/pyproject.toml
# 2
[project]
name = "OuterRail"
version = "1.2.4"
description = "AIoD - RAIL. SDK for the RAIL service of the AI on Demand platform."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "Jozef Barut",  email="jozef.barut@intern.kinit.sk"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "urllib3>=1.25.3",
    "python-dateutil>=2.8.2",
    "pydantic>=2",
    "typing-extensions>=4.7.1",
    "python-keycloak",
    "pyjwt",
    "tomlkit"
]
keywords = ["SDK", "AIoD", "RAIL", "Machine Learning", "ML Experiments"]

[project.urls]
repository = "https://github.com/aiondemand/aiod-rail/tree/feature/outer-sdk/sdk-py"
issues  = "https://github.com/aiondemand/aiod-rail/issues"
aIoD-platform = "https://www.ai4europe.eu"
RAIL-platform = "https://rail.aiod.i3a.es/docs/about>"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[tool.poetry.dev-dependencies]
pytest = ">=7.2.1"
tox = ">=3.9.0"
flake8 = ">=4.0.0"
types-python-dateutil = ">=2.8.19.14"
mypy = "1.4.1"


[tool.pylint.'MESSAGES CONTROL']
extension-pkg-whitelist = "pydantic"


[tool.setuptools.packages.find]
include = ["OuterRail*"]
namespaces = false


[tool.mypy]
files = [ "outerRail" ]



strict = true
# List from: https://mypy.readthedocs.io/en/stable/existing_code.html#introduce-stricter-options
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true

## Getting these passing should be easy
strict_equality = true
strict_concatenate = true

## Strongly recommend enabling this one as soon as you can
check_untyped_defs = true

## These shouldn't be too much additional work, but may be tricky to
## get passing if you use a lot of untyped libraries
disallow_subclassing_any = true
disallow_untyped_decorators = true
disallow_any_generics = true

## These next few are various gradations of forcing use of type annotations
disallow_untyped_calls = true
disallow_incomplete_defs = true
disallow_untyped_defs = true

## This one isn't too hard to get passing, but return on investment is lower
no_implicit_reexport = true

## This one can be tricky to get passing if you use a lot of untyped libraries
warn_return_any = true
